remove gscale2 & memory optimizations

+++ src/brightness.c
@@ -2,7 +2,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-int calcBrightness(Pixel_t *pixel) {
+unsigned short calcBrightness(Pixel_t *pixel) {
   float average;
   average = (float)(pixel->r + pixel->g + pixel->b) / 3.0f;
   return floorf(average);
@@ -14,10 +14,12 @@ Brightness_Matrix_t *getBrightness_Matrix(BitMap_t *bitmap) {
   br_matrix->width = bitmap->bmp_config->width;
   br_matrix->height = bitmap->bmp_config->height;
 
-  br_matrix->matrix = (int **)malloc(sizeof(int *) * br_matrix->height);
+  br_matrix->matrix =
+      (unsigned short **)malloc(sizeof(unsigned short *) * br_matrix->height);
 
   for (unsigned int i = 0; i < br_matrix->height; i++) {
-    br_matrix->matrix[i] = (int *)malloc(sizeof(int) * br_matrix->width);
+    br_matrix->matrix[i] =
+        (unsigned short *)malloc(sizeof(unsigned short) * br_matrix->width);
   }
 
   for (unsigned int i = 0; i < br_matrix->height; i++) {
